atualizando favicon e multer como middleware
