+++ routes/index.js
@@ -4,21 +4,7 @@ const IndexController = require("../controllers/IndexController");
 const userController = require('../controllers/userController');
 const CompraController = require('../controllers/CompraController');
 var auth = require ('../middlewares/auth');
-const multer = require('multer');
-const path = require('path')
-
-// configuração do Multer:
-const storageFile = multer.diskStorage({
-    destination: function (req, file, cb) {
-      cb(null, path.resolve(__dirname, '../public/curriculos'))
-    },
-    filename: function (req, file, cb) {
-      const uniqueSuffix = Date.now() + '.pdf'
-      cb(null, file.fieldname + '-' + uniqueSuffix)
-    }
-});
-  
-const uploadFile = multer({ storage: storageFile });
+const uploadFile = require('../middlewares/multerConfig')
 
 /* GET home page. */
 router.get('/', IndexController.index);
