+++ views/partials/header.ejs
@@ -5,6 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title><%= title %></title>
+    <link rel="shortcut icon" href="/icon/favicon.png">
     <link rel='stylesheet' href='/stylesheets/style.css' />
   </head>
